Fixed a memory problem with malloc.

+++ trunk/Stage3/DSMCBE/dsmcbe_spu_threads.h
@@ -7,8 +7,8 @@ extern int YieldThread(void);
 extern int IsThreaded(void);
 
 extern void thread_free(void* data);
-extern void* thread_malloc(unsigned int size);
+extern void* thread_malloc(unsigned long size);
 extern void thread_free_align(void* data);
-extern void* thread_malloc_align(unsigned int size, int base);
+extern void* thread_malloc_align(unsigned long size, unsigned int base);
 
 #endif /*DSMCBE_SPU_THREADS_H_*/
